requirements.txt
streamlit==1.29.0
pandas==2.1.4
plotly==5.18.0
requests==2.31.0
python-dotenv==1.0.0
jira==3.6.0
openpyxl==3.1.2